+++ zetaclient/config/config.go
@@ -4,13 +4,11 @@ package config
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
 	"sync"
 
-	"github.com/asaskevich/govalidator"
 	"github.com/fsnotify/fsnotify"
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
@@ -31,8 +29,8 @@ const folder string = "config"
 // Save saves ZetaClient config
 func Save(config *Config, path string) error {
 	// validate config
-	if err := Validate(*config); err != nil {
-		return err
+	if err := config.Validate(); err != nil {
+		return errors.Wrapf(err, "config file validation failed")
 	}
 
 	folderPath := filepath.Join(path, folder)
@@ -89,31 +87,11 @@ func Load(basePath string) (Config, error) {
 	cfg.ZetaCoreHome = basePath
 
 	// validate config
-	if err := Validate(cfg); err != nil {
-		return Config{}, err
+	if err := cfg.Validate(); err != nil {
+		return Config{}, errors.Wrapf(err, "config file validation failed")
 	}
 
 	return cfg, nil
-}
-
-// Validate performs basic validation on the config fields
-// TODO: add more validation for other fields
-// https://github.com/zeta-chain/node/issues/4352
-func Validate(cfg Config) error {
-	// go-tss requires a valid IPv4 address
-	if cfg.PublicIP != "" && !govalidator.IsIPv4(cfg.PublicIP) {
-		return fmt.Errorf("invalid public IP %s", cfg.PublicIP)
-	}
-
-	if cfg.PublicDNS != "" && !govalidator.IsDNSName(cfg.PublicDNS) {
-		return fmt.Errorf("invalid public DNS %s", cfg.PublicDNS)
-	}
-
-	if cfg.KeyringBackend != KeyringBackendFile && cfg.KeyringBackend != KeyringBackendTest {
-		return fmt.Errorf("invalid keyring backend %s", cfg.KeyringBackend)
-	}
-
-	return nil
 }
 
 // SetRestrictedAddressesFromConfig loads compliance data (restricted addresses) from config.